gcp/requirements.txt
tensorflow==2.13.1
google-cloud-storage==2.14.0
pillow==10.2.0